Add dist folder files generated with webpack
