+++ package.json
@@ -4,7 +4,7 @@
   "private": true,
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "webpack serve --open",
+    "start": "webpack serve --mode development",
     "build": "webpack"
   },
   "keywords": [],
