+++ src/index.html
@@ -6,15 +6,18 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="style.css">
     <script src="https://kit.fontawesome.com/1bc3a7e061.js" crossorigin="anonymous"></script>
-    <title>Document</title>
+    <title>To Do List</title>
 </head>
 <body>
 
 
     <div class="container">
+        <div class="title">
+            <span>Today's To Do</span> <i class="fa-solid fa-arrows-rotate"></i>
+        </div>
         <div class="form-field">
             <form action="submit">
-                <input type="text" id="input-task" class="input-task" placeholder="Add your task here...">
+                <input type="text" id="input-task" class="input-task" placeholder="Add your task here..." autocomplete="off">
                 <input type="button" id="addBtn" value="+">
             </form>
         </div>
@@ -22,10 +25,9 @@
             <ul class="toDoList" id="toDoList">
 
             </ul>
+            <input type="button" class="clearCompleted" value="Clear all completed">
         </div>
     </div>
-    
-
-    <script src="./index.js"></script>
+    <script src="index.js"></script>
 </body>
 </html>
