+++ src/index.js
@@ -1,30 +1,4 @@
-const tasks = [
-  {
-    description: 'Do exercise',
-    completed: true,
-    index: 1,
-  },
-  {
-    description: 'Practice JavaScript',
-    completed: false,
-    index: 2,
-  },
-  {
-    description: 'Make food',
-    completed: true,
-    index: 3,
-  },
-];
+import './style.css';
+import displayTasks from './modules/addToDo.js';
 
-const ulList = document.getElementById('toDoList');
-
-class displayTasks {
-  static addTask() {
-    for (let i = 0; i < tasks.length; i += 1) {
-      const task = document.createElement('li');
-      task.innerHTML = `<input type="checkbox" name="completed" id="completed"> ${tasks[i].description} <i class="fa-solid fa-ellipsis-vertical"></i>`;
-      ulList.appendChild(task);
-    }
-  }
-}
 displayTasks.addTask();
